# src/main/resources/REST-API-DOC-swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8443'
basePath: /
...
paths:
...
  /cro/stakeholders:
    get:
      tags:
        - Company Registrations Office REST API Interface
      summary: Search in or get a list of all stakeholders stored in application.
      description: Retrieve all stakeholders registered in CRO.
      operationId: searchStakeholdersUsingGET
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: StakeholderName
          in: query
          description: FirstName or LastName of the stakeholder
          required: false
          type: string
        - name: LastName
          in: query
          description: LastName of the stakeholder
          required: false
          type: string
        - name: FirstName
          in: query
          description: FirstName of the stakeholder
          required: false
          type: string
        - name: showall
          in: query
          description: Show stakeholders with all companies owned details.
          required: false
          type: boolean
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Stakeholder'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Company Registrations Office REST API Interface
      summary: Receive a Stakeholder details as a request to register it at CRO.
      description: Receive a Stakeholder details as a request to register it at CRO.
      operationId: registerStakeholderUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/CompanyRegistrationRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Stakeholder'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/cro/stakeholders/{id}':
    put:
      tags:
        - Company Registrations Office REST API Interface
      summary: Receive a Stakeholder details to update a Stakeholder on CRO.
      description: Receive an updateRequest for update an existing Stakeholder in CRO.
      operationId: updateStakeholderUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
        - in: body
          name: updateRequest
          description: updateRequest
          required: true
          schema:
            $ref: '#/definitions/CompanyRegistrationRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Stakeholder'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/cro/stakeholders/{stakeholderId}':
    get:
      tags:
        - Company Registrations Office REST API Interface
      summary: Search a stakeholder based on the received stakeholder name.
      description: Search a stakeholder based on the received stakeholder name.
      operationId: getStakeholderByIdUsingGET
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: stakeholderId
          in: path
          description: stakeholderId
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Stakeholder'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
...
  Company:
    type: object
    required:
      - Stakeholders
    properties:
      Address:
        type: string
      City:
        type: string
      Country:
        type: string
      Email:
        type: string
      Id:
        type: integer
        format: int32
      Name:
        type: string
      PhoneNumber:
        type: string
      Stakeholders:
        type: array
        description: A collection of the returned company stakeholders
        allowEmptyValue: false
        items:
          $ref: '#/definitions/Stakeholder'
    title: Company
    description: Company resource representation
  CompanyRegistrationRequest:
    type: object
    properties:
      FirstName:
        type: string
      LastName:
        type: string
    title: CompanyRegistrationRequest
    description: Create Stakeholder registration request resource representation
...
  Stakeholder:
    type: object
    required:
      - Companies
    properties:
      Companies:
        type: array
        description: A list of companies details owned by this stakeholder
        allowEmptyValue: false
        items:
          $ref: '#/definitions/Company'
      FirstName:
        type: string
      Id:
        type: integer
        format: int32
      LastName:
        type: string
    title: Stakeholder
    description: Stakeholder resource representation
...